consolidate tracing/ and _trace/ subdirectories because they serve basically the same purpose
