+++ ddtrace/contrib/__init__.py
@@ -1,4 +1,4 @@
+from .._trace import trace_handlers  # noqa:F401
 from ..internal.utils.importlib import func_name  # noqa:F401
 from ..internal.utils.importlib import module_name  # noqa:F401
 from ..internal.utils.importlib import require_modules  # noqa:F401
-from ..tracing import trace_handlers  # noqa:F401
